requirements.txt
markdown
Pillow